take entire committee member list section to committee component
